Add VSAC Base class
