+++ services/visual-studio-app-center/visual-studio-app-center-releases-size.service.js
@@ -1,23 +1,17 @@
 'use strict'
 
 const Joi = require('@hapi/joi')
-const { BaseJsonService } = require('..')
+const {
+  BaseVisualStudioAppCenterService,
+  keywords,
+  documentation,
+} = require('./visual-studio-app-center-base')
 
 const schema = Joi.object({
   size: Joi.number().required(),
 }).required()
 
-const keywords = [
-  'visual-studio',
-  'vsac',
-  'visual-studio-app-center',
-  'app-center',
-]
-
-const documentation =
-  "You will need to create a <b>read-only</b> API token <a target='_blank' href='https://appcenter.ms/settings/apitokens'>here</a>."
-
-module.exports = class VisualStudioAppCenterReleasesSize extends BaseJsonService {
+module.exports = class VisualStudioAppCenterReleasesSize extends BaseVisualStudioAppCenterService {
   static get category() {
     return 'size'
   }
@@ -58,27 +52,8 @@ module.exports = class VisualStudioAppCenterReleasesSize extends BaseJsonService
     }
   }
 
-  async fetch({ owner, app, token }) {
-    const url = `https://api.appcenter.ms/v0.1/apps/${owner}/${app}/releases/latest`
-
-    return this._requestJson({
-      schema,
-      options: {
-        headers: {
-          'X-API-Token': token,
-        },
-      },
-      errorMessages: {
-        401: 'invalid token',
-        403: 'project not found',
-        404: 'project not found',
-      },
-      url,
-    })
-  }
-
-  async handle({ owner, app, branch, token }) {
-    const json = await this.fetch({ owner, app, branch, token })
+  async handle({ owner, app, token }) {
+    const json = await this.fetch({ owner, app, token, schema })
     return this.constructor.render({ size: json.size })
   }
 }
